dom/base/test/test_window_constructor.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=824217
-->
<head>
  <meta charset="UTF-8">
  <title>Test for Bug 824217</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=622491">Mozilla Bug 824217</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<iframe name="constructor" src="about:blank"></iframe>
<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

function run()
{
  // Ideally we'd test that this property lives on the right place in the
  // [[Prototype]] chain, with the right attributes there, but we don't
  // implement this right yet, so just test the value's what's expected.
  is(window.constructor, Window,
     "should have gotten Window, not the constructor frame");
  SimpleTest.finish();
}

window.addEventListener("load", run, false);
</script>
</pre>
</body>
</html>
